requirements.txt
pygame
numpy
inferactively-pymdp

